requirements.txt
requests==2.14.2
pillow==6.2.1
urllib3==1.25.10